Move ScimUser.Group into separate class.
