+++ scim/src/main/java/org/cloudfoundry/identity/uaa/scim/ScimUser.java
@@ -40,96 +40,6 @@ import org.springframework.util.StringUtils;
 public final class ScimUser extends ScimCore {
 
     @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
-    public static final class Group {
-        String value;
-
-        String display;
-
-        public static enum Type {
-            DIRECT, INDIRECT
-        };
-
-        Type type;
-
-        public Type getType() {
-            return type;
-        }
-
-        public void setType(Type type) {
-            this.type = type;
-        }
-
-        public Group() {
-            this(null, null);
-        }
-
-        public Group(String value, String display) {
-            this(value, display, Type.DIRECT);
-        }
-
-        public Group(String value, String display, Type type) {
-            this.value = value;
-            this.display = display;
-            this.type = type;
-        }
-
-        public String getValue() {
-            return value;
-        }
-
-        public void setValue(String value) {
-            this.value = value;
-        }
-
-        public String getDisplay() {
-            return display;
-        }
-
-        public void setDisplay(String display) {
-            this.display = display;
-        }
-
-        @Override
-        public int hashCode() {
-            final int prime = 31;
-            int result = 1;
-            result = prime * result + ((display == null) ? 0 : display.hashCode());
-            result = prime * result + ((value == null) ? 0 : value.hashCode());
-            result = prime * result + ((type == null) ? 0 : type.hashCode());
-            return result;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj)
-                return true;
-            if (obj == null)
-                return false;
-            if (getClass() != obj.getClass())
-                return false;
-            Group other = (Group) obj;
-            if (display == null) {
-                if (other.display != null)
-                    return false;
-            }
-            else if (!display.equals(other.display))
-                return false;
-            if (value == null) {
-                if (other.value != null)
-                    return false;
-            }
-            else if (!value.equals(other.value))
-                return false;
-            return type == other.type;
-        }
-
-        @Override
-        public String toString() {
-            return String.format("(id: %s, name: %s, type: %s)", value, display, type);
-        }
-    }
-
-    @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
     public static final class Name {
         String formatted;
 
@@ -208,7 +118,7 @@ public final class ScimUser extends ScimCore {
 
     private List<ScimEmail> emails;
 
-    private Set<Group> groups;
+    private Set<ScimUserGroup> groups;
 
     private Set<Approval> approvals;
 
@@ -286,12 +196,12 @@ public final class ScimUser extends ScimCore {
         this.approvals = approvals;
     }
 
-    public Set<Group> getGroups() {
+    public Set<ScimUserGroup> getGroups() {
         return groups;
     }
 
-    public void setGroups(Collection<Group> groups) {
-        this.groups = new LinkedHashSet<Group>(groups);
+    public void setGroups(Collection<ScimUserGroup> groups) {
+        this.groups = new LinkedHashSet<ScimUserGroup>(groups);
     }
 
     public List<ScimPhoneNumber> getPhoneNumbers() {
