+++ uaa/src/test/java/org/cloudfoundry/identity/uaa/integration/VmcUserIdTranslationEndpointIntegrationTests.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- *     Cloud Foundry 
+ *     Cloud Foundry
  *     Copyright (c) [2009-2014] Pivotal Software, Inc. All Rights Reserved.
  *
  *     This product is licensed to you under the Apache License, Version 2.0 (the "License").
@@ -21,7 +21,7 @@ import java.util.Map;
 
 import org.cloudfoundry.identity.uaa.message.PasswordChangeRequest;
 import org.cloudfoundry.identity.uaa.scim.ScimUser;
-import org.cloudfoundry.identity.uaa.scim.ScimUser.Group;
+import org.cloudfoundry.identity.uaa.scim.ScimUserGroup;
 import org.cloudfoundry.identity.uaa.test.TestAccountSetup;
 import org.cloudfoundry.identity.uaa.test.UaaTestAccounts;
 import org.junit.Assume;
@@ -41,7 +41,7 @@ import org.springframework.web.client.RestOperations;
 /**
  * Integration test to verify that the userid translation use cases are
  * supported adequately for vmc.
- * 
+ *
  * @author Luke Taylor
  */
 @OAuth2ContextConfiguration(OAuth2ContextConfiguration.Implicit.class)
@@ -80,7 +80,7 @@ public class VmcUserIdTranslationEndpointIntegrationTests {
         user.setUserName(JOE);
         user.setName(new ScimUser.Name("Joe", "User"));
         user.addEmail("joe@blah.com");
-        user.setGroups(Arrays.asList(new Group(null, "uaa.user"), new Group(null, "orgs.foo")));
+        user.setGroups(Arrays.asList(new ScimUserGroup(null, "uaa.user"), new ScimUserGroup(null, "orgs.foo")));
 
         ResponseEntity<ScimUser> newuser = client.postForEntity(serverRunning.getUrl(userEndpoint), user,
                         ScimUser.class);
